backend/save_qualification/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
requests>=2.28
